+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # SecretSauce
+  * 0.5.1 implementing `NavigationDrawerFragment.NavigationDrawerCallbacks` is now optional for `BaseAppCompatActivty`. 
+    Deprecated activities removed.
+    Methods for showing non support fragments removed from `BaseAppCompatActivty`
+    Support for `WaitDialog` removed from `BaseAppCompatActivty`. Use `WaitDialogHandler` in your activity if you need it.
   * 0.5.0 `getContainerId` and `getDefaultFontName` moved from `BaseApp` to `AppComponentInterface`
   * 0.4.9 `@Named(BaseApp.INJECT_NAME_SESSION_ID)` replaced with `@SessionId`;
     `CheckableFrameLayout` and `ChackableLinearLayout` updated to pass real view instead of null;